SVM: allow direct use of system allocator buffers (experimental)
Use shared system USM aka SVM if it is available on Intel GPU. Instead of
being enabled by default, it would be best to only use it if requested by
the user explicitely with a new flag.

For now it is only enabled in the benchmarks therefore only in the memory
allocations under "defines" used by the benchmark. Early results show the
performance impact is limited to <1% when using SVM for "rho" and "u" but
the degradation increases when it is extended to "fi".

Signed-off-by: Francois Dugast <francois.dugast@intel.com>